blog delete

+++ admin/view_b_post.php
@@ -85,7 +85,7 @@
                                         <a href="blog_details.php?id=<?php echo $row->blog_post_id; ?>" class="btn btn-success btn-sm">View</a>
                                     </td>
                                     <td>
-                                        <a href="blog_delete.php?id=<?php echo $row->blog_post_id; ?>" class="btn btn-danger btn-sm">Delete</a>
+                                        <a onclick="javascript:return confirm('Are You Sure?')" href="blog_delete.php?id=<?php echo $row->blog_post_id; ?>" class="btn btn-danger btn-sm">Delete</a>
                                     </td>
                                 </tr>
                             <?php
